FakeDeafen update, and reseat philhks plugins lol
FakeDeafen now has a button next to the user panel.
